+++ backend/src/main/java/com/archivo/backend/repositories/SedeRepository.java
@@ -1,10 +1,12 @@
 package com.archivo.backend.repositories;
 
 import com.archivo.backend.entities.Sede;
-import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface SedeRepository extends JpaRepository<Sede, Long> {
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SedeRepository extends JpaRepository<Sede, Integer> {
+    // Puede añadir métodos de búsqueda personalizados aquí si es necesario
     Optional<Sede> findByNombre(String nombre);
 }
